chore(sentry): split suite-common/sentry files
