+++ suite-common/sentry/src/constants.ts
@@ -1,0 +1,3 @@
+export const allowReportTag = 'allowReport';
+export const coinjoinReportTag = 'coinjoinReport';
+export const coinjoinNetworkTag = 'coinjoinNetwork';
